organized components folder and sass
