+++ src/ScssComponends/_budget-cards.scss
@@ -2,10 +2,15 @@
     border: 1px solid black;
     width: 500px;
     display: flex;
+    justify-content: center;
 }
 .titles{
     display: flex;
     flex-direction: row;
     justify-content: space-between;
     padding: 1%;
+    width: 90%;
+}
+.max{
+    font-size: 0.9rem;
 }
